added component UpdateOrder for logged users
