+++ app/(routes)/(auth)/user-panel/page.tsx
@@ -3,32 +3,32 @@ import {
   addProduct,
   clearCart,
 } from "@/app/GlobalRedux/Features/cart/cartSlice";
-import { RootState } from "@/app/GlobalRedux/store";
 import { AddedProduct } from "@/app/types/types";
 import { signOut } from "next-auth/react";
 import { useSession } from "next-auth/react";
-import { useRouter } from "next/router";
 import { useEffect, useCallback } from "react";
-import { useDispatch, useSelector } from "react-redux";
+import { useDispatch } from "react-redux";
 
 const Page = () => {
   const session: any = useSession();
   const dispatch = useDispatch();
-  const cartProducts = useSelector((state: RootState) => state.products);
 
   const getCart = useCallback(async () => {
-    const res = await fetch("/api/getUserCart", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ id: session.id }),
-    });
-    console.log(res);
-    if (res.ok && !localStorage.getItem("cart")) {
-      dispatch(clearCart());
-      const { cart } = await res.json();
-      cart.forEach((product: AddedProduct) => dispatch(addProduct(product)));
+    try {
+      const res = await fetch("/api/getUserCart", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ id: session.id }),
+      });
+      if (res.ok && !localStorage.getItem("cart")) {
+        dispatch(clearCart());
+        const { cart } = await res.json();
+        cart.forEach((product: AddedProduct) => dispatch(addProduct(product)));
+      }
+    } catch (error) {
+      console.log("error on gettingcart from db");
     }
   }, [session.id, dispatch]);
